merge: move temporaldata to torch_brain/data + tests
+ rm dependencies: torch and pydantic
