--- temporaldata/tests/test_version.py
@@ -1,6 +1,0 @@
-import temporaldata
-
-
-def test_version_exists():
-    assert hasattr(temporaldata, "__version__")
-    assert len(temporaldata.__version__) > 0
