+++ tests/data/test_data.py
@@ -7,7 +7,7 @@ import h5py
 import numpy as np
 import pytest
 
-from temporaldata import (
+from torch_brain.data import (
     ArrayDict,
     Data,
     Interval,
