+++ torch_brain/__init__.py
@@ -1,10 +1,4 @@
 from importlib.metadata import version, PackageNotFoundError
-from . import data
-from . import nn
-from . import models
-from . import utils
-from . import transforms
-from . import samplers
 
 try:
     __version__ = version("torch_brain")
