+++ torch_brain/data/__init__.py
@@ -1,18 +1,31 @@
-from . import sampler
-from . import collate
-from .dataset import Dataset
+from importlib.metadata import version, PackageNotFoundError
+from .arraydict import ArrayDict, LazyArrayDict
+from .irregular_ts import IrregularTimeSeries, LazyIrregularTimeSeries
+from .regular_ts import RegularTimeSeries, LazyRegularTimeSeries
+from .interval import Interval, LazyInterval
+from .data import Data
 
-__all__ = ["Dataset"]
+from .concat import concat
 
-# see docs/source/api_reference.py
+__all__ = [
+    "ArrayDict",
+    "LazyArrayDict",
+    "IrregularTimeSeries",
+    "LazyIrregularTimeSeries",
+    "RegularTimeSeries",
+    "LazyRegularTimeSeries",
+    "Interval",
+    "LazyInterval",
+    "Data",
+    "concat",
+]
+
+# Drives the generated API reference; see docs/source/api_reference.py.
 __api_ref__ = {
     "description": None,
     "sections": [
         {
-            "title": "torch_brain.data.dataset [DEPRECATED]",
-            "autosummary": [
-                "Dataset",
-            ],
+            "autosummary": __all__,
         },
     ],
 }
