+++ torch_brain/data/regular_ts.py
